Replicator: Don't fetch remote checkpoint if there's no local checkpoint
Reduces replication time by eliminating an HTTP round-trip delay on an
initial replication.

I had to handle the case where the local db has no saved checkpoint but
has been copied from a pre-installed database; in that case we do need
to use the regular code path.

Also added a bit of code in CBLChangeTracker to map a low-level DNS
error to a more usual NSURL error; this was to keep a unit test working.
When we skip the checkpoint lookup, and the remote hostname is bogus,
the error that stops replication will come from the change tracker not
a RemoteRequest, so its error domain/code were different.