Partially fixed CBLWebSocketChangeTracker unzip errors
I still don't know why the first error occurs. But after that, the
ChangeTracker wouldn't be able to recover, because it would reconnect
but keep using the same GZip decoder, which wouldn't be in an initial
state. So it would keep getting gzip errors every time. I fixed this
by nil-ing out _gzip in the -start method.

I also improved the flow of control so, after it gets an error and
decides to close the WebSocket, it will ignore subsequent queued
messages to avoid getting confused by them and logging more errors.

For #1074