Fixed race causing the change tracker to start twice
The issue couldn’t reproduce but here is what is speculated to happen:

There could be a race scenario between stopping the fetch remote checkpoint request when putting the replicator offine and the request itself is completed. If stopping the request is too late, the onComplete will be called in the next runloop, and that will cause a change tracker to start in offline mode.

To prevent above scenario, when the request is completed, check whether the replicator is still online or not. If the replicator is not online, just finish the request and return.

#1655